requirements.txt
flask
pandas
matplotlib
seaborn
openai
python-dotenv
